Se refactoriza codigo creando custom hooks
