+++ src/lib/navigation.js
@@ -1,0 +1,26 @@
+export const navigation = [
+  {
+    name: "My Petit",
+    href: "/petit",
+    current: false,
+    private: true,
+  },
+  {
+    name: "Reportes",
+    href: "/reports",
+    current: false,
+    private: true,
+  },
+  {
+    name: "Acerca de Nosotros",
+    href: "/about",
+    current: false,
+    private: false,
+  },
+  {
+    name: "Notifications",
+    select: true,
+    current: false,
+    private: true,
+  },
+];
